requirements.txt
42cc-pystyle==0.0.15
Django==1.6.7
South==1.0.2
argparse==1.2.1
django-bootstrap3==6.2.2
flake8==2.5.2
mccabe==0.4.0
pep8==1.7.0
pyflakes==1.0.0
wsgiref==0.1.2
Pillow==3.1.0
